move TransactionContext from js to tracing
